requirements.txt
tensorflow==2.15.0
numpy==1.24.3
opencv-python==4.8.1.78
matplotlib==3.8.0
scikit-learn==1.3.2
tqdm==4.66.1
Pillow==10.1.0
